bench: split Blake3 nonregression runner
